requirements.txt
flask==2.3.3
flask-sqlalchemy==3.0.5
flask-wtf==1.1.1
flask-login==0.6.3
email-validator==2.0.0
pillow==10.0.0
bcrypt==4.0.1
wtforms==3.0.1